Added product update method

+++ src/main/java/com/demo/store/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.demo.store.repositories;
 import com.demo.store.entities.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Byte> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
 }
